+++ character.cc
@@ -1,4 +1,3 @@
-#include "character.h"
 #include <iostream>
 #include <map>
 #include <string>