+++ src/npc.cc
@@ -1,44 +1,42 @@
-#include <iostream>
-#include <string>
-#include <map>
-#include <vector>
+#include "npc.h"
 #include "character.h"
+#include <iostream>
+#include <map>
+#include <string>
+#include <vector>
 
-class NPC : public Character {
-    private:
-        std::string type;
-    public:
-        int loc_id;
-        std::vector<bool> interactions;
+//***CONSTRUCTORS***//
+// Default constructor
+NPC::NPC() : loc_id(0), type("Unknown"), interactions({false}) {}
 
-        // Nearly-default constructor
-        NPC(std::string name_) : Character(name_) {
-            loc_id = 0;
-            interactions.push_back(false);
-            type = "Unknown";
-        }
+// Nearly-default constructor
+NPC::NPC(std::string name_)
+    : Character(name_), type("Unknown"), loc_id(0), interactions({false}) {}
 
-        // Construct generic NPC based on type
-        NPC(std::string name_, std::string type_) : Character(name_) {
-            type = type_;
+// Construct generic NPC based on type
+NPC::NPC(std::string name_, std::string type_, int loc_id_)
+    : Character(name_), type(type_), loc_id(loc_id_) {
 
-            // TODO: generic Character archetypes
-            if(type == "farmer") {
-            } else if (type == "warrior") {
-            } else if (type == "rogue") {
-            } else if (type == "priest") {
-            } else if (type == "mage") {
-            } else if (type == "animal") {
-            } else {}
-        }
+  // TODO: generic Character archetypes
+  if (type == "farmer") {
+  } else if (type == "warrior") {
+  } else if (type == "rogue") {
+  } else if (type == "priest") {
+  } else if (type == "mage") {
+  } else if (type == "animal") {
+  } else {
+  }
+}
 
-        NPC(Character& c, std::string type_, int loc_id_,
-        std::vector<bool> interactions_) : Character(c) {
-                type = type_;
-                loc_id = loc_id_;
-                interactions = interactions_;
-            }
+// Full parameterized constructor
+NPC::NPC(Character &c, std::string type_, int loc_id_,
+         std::vector<bool> interactions_)
+    : Character::Character(c), type(type_), loc_id(loc_id_),
+      interactions(interactions_) {}
 
-        // Overload Character print() method to include NPC information
-        std::string print();
-};
+//***METHODS***//
+// type getter()
+std::string NPC::get_type() { return type; }
+
+// Overload Character print() method to include NPC information
+std::string NPC::print() { return ""; }